third_party/blink/web_tests/external/wpt/uievents/mouse/mousemove-action-interpolation-expected.txt

This is a testharness.js-based test.
[FAIL] Mousemove should transition smoothly over specified duration
  assert_greater_than: Should have more than 10 mousemove events over 500ms expected a number greater than 10 but got 2
Harness: the test ran to completion.
